Added checking info file

+++ ai_video_maker/story_files/make_story_files.py
@@ -1,5 +1,6 @@
 import os
 import re
+import json
 from ai_video_maker.directory import change_dir_name
 from ai_video_maker.statuses import change_status_available_files
 
@@ -13,23 +14,34 @@ def make_story_files(reddit):
     # Get Title and selfText by ID's & put into file texts
     for directory in os.listdir("./stories"):
         if os.path.isdir(f"stories/{directory}"):
-            with open("stories_id/stories_id.txt", "r+") as stories_id:
-                first_line = stories_id.readline()
-                lines = stories_id.readlines()
-
-                with open("stories_id/used_stories_id.txt", "a") as used_stories_id:
-                    used_stories_id.write(first_line)
-                file_name = int(re.sub("[a-z]|[A-Z]", "", directory))
-
+            if os.path.isdir(f"stories/{directory}"):
                 with open(
-                    f"stories/{directory}/story{file_name:04d}.txt", "wb"
-                ) as story_file:
-                    output = f"{reddit.submission(id=first_line).title}\n\n{reddit.submission(id=first_line).selftext}"
+                    f"stories/{directory}/info.json", "r", encoding="utf-8"
+                ) as file:
+                    data = json.load(file)
 
-                    output = output.replace("AITA", "Am I The Asshole")
-                    story_file.write(output.encode("utf-8"))
+                    if isinstance(data, dict):
+                        if data.get("status") == "Empty":
+                            with open("stories_id/stories_id.txt", "r+") as stories_id:
+                                first_line = stories_id.readline()
+                                lines = stories_id.readlines()
 
-                    stories_id.seek(0)
-                    stories_id.truncate()
-                    stories_id.writelines(lines[1:])
-            change_status_available_files(directory, "story file")
+                                with open(
+                                    "stories_id/used_stories_id.txt", "a"
+                                ) as used_stories_id:
+                                    used_stories_id.write(first_line)
+                                file_name = int(re.sub("[a-z]|[A-Z]", "", directory))
+
+                                with open(
+                                    f"stories/{directory}/story{file_name:04d}.txt",
+                                    "wb",
+                                ) as story_file:
+                                    output = f"{reddit.submission(id=first_line).title}\n\n{reddit.submission(id=first_line).selftext}"
+
+                                    output = output.replace("AITA", "Am I The Asshole")
+                                    story_file.write(output.encode("utf-8"))
+
+                                    stories_id.seek(0)
+                                    stories_id.truncate()
+                                    stories_id.writelines(lines[1:])
+                            change_status_available_files(directory, "story file")
